refactor(modeling): moved retessellate functions to modifiers
