+++ packages/modeling/src/operations/booleans/unionGeom3.js
@@ -1,6 +1,7 @@
 const flatten = require('../../utils/flatten')
 
-const retessellate = require('./retessellate')
+const retessellate = require('../modifiers/retessellate')
+
 const unionSub = require('./unionGeom3Sub')
 
 /*
